requirements.txt
black==22.1.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
Flask-Mail==0.9.1
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
python-dotenv==0.19.2
qrcode==7.3.1
requests==2.27.1
tomli==2.0.1
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
